Fixes #142 exit code sometimes changed
Gating the new output now on args.raw instead of the return value
which could sometimes return things like [] which sys.exit (specifically
_Py_HandleSystemExit) would return a 1 on.

Now the behavior is if args.raw was specified, it exits(0) which of
course could eventually have other potential side effects for exception
path handling.

Regardless, that seems to not currently be a problem and quick fixes are
nice.